app/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

confluent-kafka==2.11.1

pymongo==4.6.3
sqlalchemy==2.0.30
mysql-connector-python==8.4.0

requests==2.31.0
python-dotenv==1.0.1
